requirements.txt
discord.py
aiohttp
aiofiles
ruamel.yaml
pyyaml
colorama
requests
packaging
PyNaCl
openai
tiktoken
psutil
ollama
anthropic